Merge master in staging

+++ core/api/offers/server/hooks.js
@@ -6,6 +6,4 @@ Offers.before.remove((userId, { _id: offerId }) => {
   OfferService.cleanUpOffer({ offerId });
 });
 
-
-Offers.after.remove((userId, { _id }) =>
-  FileService.deleteAllFilesForDoc(_id));
+Offers.after.remove((userId, { _id }) => FileService.deleteAllFilesForDoc(_id));
